+++ thoth-app/src/component/catalogue.rs
@@ -1,26 +1,47 @@
-use std::str::FromStr;
 use yew::html;
-use yew::prelude::*;
+use yew::prelude::Component;
+use yew::prelude::FocusEvent;
+use yew::prelude::Html;
+use yew::prelude::InputData;
+use yew::prelude::ShouldRender;
 use yew::ComponentLink;
+use yewtil::fetch::Fetch;
 use yewtil::fetch::FetchAction;
 use yewtil::fetch::FetchState;
 use yewtil::future::LinkFuture;
 
+use crate::component::utils::Loader;
 use crate::component::utils::Reloader;
 use crate::models::work::works_query::FetchActionWorks;
 use crate::models::work::works_query::FetchWorks;
-use crate::models::work::License;
+use crate::models::work::works_query::Variables;
+use crate::models::work::works_query::WorksRequest;
+use crate::models::work::works_query::WorksRequestBody;
 use crate::models::work::Work;
-use crate::THOTH_API;
 
 pub struct CatalogueComponent {
-    markdown: FetchWorks,
+    limit: i32,
+    offset: i32,
+    page_size: i32,
+    search_term: String,
+    data: Vec<Work>,
+    result_count: i32,
+    fetch_data: FetchWorks,
     link: ComponentLink<Self>,
 }
 
+pagination_helpers! {CatalogueComponent, PAGINATION_COUNT_WORKS, SEARCH_WORKS}
+
 pub enum Msg {
-    SetMarkdownFetchState(FetchActionWorks),
-    GetMarkdown,
+    SetFetchState(FetchActionWorks),
+    GetData,
+    PaginateData,
+    #[allow(dead_code)]
+    Search(String),
+    ChangeSearchTerm(String),
+    TriggerSearch,
+    NextPage,
+    PreviousPage,
 }
 
 impl Component for CatalogueComponent {
@@ -28,32 +49,90 @@ impl Component for CatalogueComponent {
     type Properties = ();
 
     fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
-        CatalogueComponent {
-            markdown: Default::default(),
-            link,
-        }
-    }
+        let offset: i32 = Default::default();
+        let page_size: i32 = 10;
+        let limit: i32 = page_size;
+        let search_term: String = Default::default();
+        let result_count: i32 = Default::default();
+        let data = Default::default();
+        let fetch_data = Default::default();
 
-    fn rendered(&mut self, first_render: bool) {
-        if first_render {
-            self.link
-                .send_future(self.markdown.fetch(Msg::SetMarkdownFetchState));
-            self.link
-                .send_message(Msg::SetMarkdownFetchState(FetchAction::Fetching));
+        link.send_message(Msg::PaginateData);
+
+        CatalogueComponent {
+            limit,
+            offset,
+            page_size,
+            search_term,
+            data,
+            result_count,
+            fetch_data,
+            link,
         }
     }
 
     fn update(&mut self, msg: Self::Message) -> ShouldRender {
         match msg {
-            Msg::SetMarkdownFetchState(fetch_state) => {
-                self.markdown.apply(fetch_state);
+            Msg::SetFetchState(fetch_state) => {
+                self.fetch_data.apply(fetch_state);
+                self.data = match self.fetch_data.as_ref().state() {
+                    FetchState::Fetched(body) => body.data.works.clone(),
+                    _ => Default::default(),
+                };
+                self.result_count = match self.fetch_data.as_ref().state() {
+                    FetchState::Fetched(body) => body.data.work_count,
+                    _ => Default::default(),
+                };
                 true
             }
-            Msg::GetMarkdown => {
+            Msg::GetData => {
                 self.link
-                    .send_future(self.markdown.fetch(Msg::SetMarkdownFetchState));
+                    .send_future(self.fetch_data.fetch(Msg::SetFetchState));
                 self.link
-                    .send_message(Msg::SetMarkdownFetchState(FetchAction::Fetching));
+                    .send_message(Msg::SetFetchState(FetchAction::Fetching));
+                false
+            }
+            Msg::PaginateData => {
+                let filter = self.search_term.clone();
+                let body = WorksRequestBody {
+                    variables: Variables {
+                        limit: Some(self.limit),
+                        offset: Some(self.offset),
+                        filter: Some(filter),
+                    },
+                    ..Default::default()
+                };
+                let request = WorksRequest { body };
+                self.fetch_data = Fetch::new(request);
+                self.link.send_message(Msg::GetData);
+                false
+            }
+            Msg::Search(_) => {
+                // needed because of macro, but unused here
+                false
+            }
+            Msg::ChangeSearchTerm(term) => {
+                self.search_term = term;
+                false
+            }
+            Msg::TriggerSearch => {
+                self.limit = self.page_size;
+                self.offset = 0;
+                self.link.send_message(Msg::PaginateData);
+                false
+            }
+            Msg::NextPage => {
+                if self.limit < self.result_count && !self.is_next_disabled() {
+                    self.offset += self.page_size;
+                    self.link.send_message(Msg::PaginateData);
+                }
+                false
+            }
+            Msg::PreviousPage => {
+                if self.offset > 0 && !self.is_previous_disabled() {
+                    self.offset -= self.page_size;
+                    self.link.send_message(Msg::PaginateData);
+                }
                 false
             }
         }
@@ -64,166 +143,70 @@ impl Component for CatalogueComponent {
     }
 
     fn view(&self) -> Html {
-        match self.markdown.as_ref().state() {
-            FetchState::NotFetching(_) => {
-                html! {<Reloader onclick=self.link.callback(|_| Msg::GetMarkdown)/>}
-            }
-            FetchState::Fetching(_) => html! {
-                <div class="pageloader is-active is-warning">
-                    <span class="title">{ "Loading" }</span>
-                 </div>
-            },
-            FetchState::Fetched(body) => html! {
-                <div class="container">
-                    { for body.data.works.iter().map(render_work) }
-                </div>
-            },
-            FetchState::Failed(_, err) => html! {&err},
-        }
-    }
-}
-
-fn render_license(license: &License) -> Html {
-    html! {
-        <span class="icon is-small license">
-            <i class="fab fa-creative-commons" aria-hidden="true"></i>
-            {
-                match license {
-                    License::By =>html!{
-                        <i class="fab fa-creative-commons-by" aria-hidden="true"></i>
-                    },
-                    License::BySa => html!{
-                        <>
-                            <i class="fab fa-creative-commons-by" aria-hidden="true"></i>
-                            <i class="fab fa-creative-commons-sa" aria-hidden="true"></i>
-                        </>
-                    },
-                    License::ByNd => html!{
-                        <>
-                            <i class="fab fa-creative-commons-by" aria-hidden="true"></i>
-                            <i class="fab fa-creative-commons-nd" aria-hidden="true"></i>
-                        </>
-                    },
-                    License::ByNc => html!{
-                        <>
-                            <i class="fab fa-creative-commons-by" aria-hidden="true"></i>
-                            <i class="fab fa-creative-commons-nc" aria-hidden="true"></i>
-                        </>
-                    },
-                    License::ByNcSa => html!{
-                        <>
-                            <i class="fab fa-creative-commons-by" aria-hidden="true"></i>
-                            <i class="fab fa-creative-commons-nc" aria-hidden="true"></i>
-                            <i class="fab fa-creative-commons-sa" aria-hidden="true"></i>
-                        </>
-                    },
-                    License::ByNcNd => html!{
-                        <>
-                            <i class="fab fa-creative-commons-by" aria-hidden="true"></i>
-                            <i class="fab fa-creative-commons-nc" aria-hidden="true"></i>
-                            <i class="fab fa-creative-commons-nd" aria-hidden="true"></i>
-                        </>
-                    },
-                    License::Zero => html!{
-                        <i class="fab fa-creative-commons-zero" aria-hidden="true"></i>
-                    },
-                    License::Undefined => html! {}
-                }
-            }
-        </span>
-    }
-}
-
-fn render_work(w: &Work) -> Html {
-    let doi = &w.doi.clone().unwrap_or_else(|| "".to_string());
-    let license = License::from_str(&w.license.clone().unwrap_or_else(|| "".to_string())).unwrap();
-    let cover_url = &w.cover_url.clone().unwrap_or_else(|| "".to_string());
-    let place = &w.place.clone().unwrap_or_else(|| "".to_string());
-    html! {
-        <div class="box">
-            <article class="media">
-                <div class="media-left">
-                <figure class="image is-96x96">
-                    <img src={cover_url} alt={format!("{} - Cover Image", &w.title)} loading="lazy" />
-                    { render_license(&license) }
-                </figure>
-                </div>
-                <div class="media-content">
-                    <div class="content">
-                        <p>
-                            <strong>{&w.full_title}</strong>
-                            <br/>
-                            <div>
-                            {
-                                if let Some(contributions) = &w.contributions {
-                                    contributions.iter().map(|c| c.main_contribution_item_bullet_small()).collect::<Html>()
-                                } else {
-                                    html! {}
-                                }
-                            }
-                            </div>
-                            <br/>
-                            {
-                                if let Some(date) = &w.publication_date {
-                                    let mut c1 = date.chars();
-                                    c1.next();
-                                    c1.next();
-                                    c1.next();
-                                    c1.next();
-                                    let year: &str = &date[..date.len() - c1.as_str().len()];
-                                    html! {<small>{place}{": "}{&w.imprint.publisher.publisher_name}{", "}{year}</small>}
-                                } else {
-                                    html! {<small>{&w.imprint.publisher.publisher_name}</small>}
-                                }
-                            }
-                            <br/>
-                            <small>{&doi}</small>
+        html! {
+            <div class="container">
+                <nav class="level">
+                    <div class="level-left">
+                        <p class="level-item">
+                            <span>
+                            { self.display_count() }
+                            </span>
                         </p>
                     </div>
-                    <nav class="level is-mobile">
-                        <div class="level-left">
-                            <a
-                                class="level-item button is-small"
-                                aria-label="read"
-                                href={format!("{}", doi)}
-                            >
-                                <span class="icon is-small">
-                                <i class="fas fa-book" aria-hidden="true"></i>
-                                </span>
-                                <span>{"Read"}</span>
-                            </a>
-
-                            <div class="level-item dropdown is-hoverable">
-                                <div class="dropdown-trigger">
-                                    <button
-                                        class="button is-small"
-                                        aria-haspopup="true"
-                                        aria-controls="dropdown-menu"
-                                    >
-                                        <span class="icon is-small">
-                                            <i class="fas fa-file" aria-hidden="true"></i>
-                                        </span>
-                                        <span>{"Metadata"} </span>
-                                        <span class="icon is-small">
-                                            <i class="fas fa-angle-down" aria-hidden="true"></i>
-                                        </span>
+                    <div class="level-right" />
+                </nav>
+                <nav class="pagination is-centered" role="navigation" aria-label="pagination">
+                    <a class="pagination-previous"
+                        onclick=self.link.callback(|_| Msg::PreviousPage)
+                        disabled=self.is_previous_disabled()
+                    >{ crate::string::PREVIOUS_PAGE_BUTTON }</a>
+                    <a class="pagination-next"
+                        onclick=self.link.callback(|_| Msg::NextPage)
+                        disabled=self.is_next_disabled()
+                    >{ crate::string::NEXT_PAGE_BUTTON }</a>
+                    <div class="pagination-list">
+                        <form
+                            style="width: 80%"
+                            onsubmit=self.link.callback(|e: FocusEvent| {
+                                e.prevent_default();
+                                Msg::TriggerSearch
+                            })
+                        >
+                            <div class="field has-addons">
+                                <p class="control is-expanded has-icons-left">
+                                    <input
+                                        class="input"
+                                        type="search"
+                                        value=self.search_term
+                                        placeholder=self.search_text()
+                                        oninput=self.link.callback(|e: InputData| Msg::ChangeSearchTerm(e.value))
+                                    />
+                                    <span class="icon is-left">
+                                        <i class="fas fa-search" aria-hidden="true"></i>
+                                    </span>
+                                </p>
+                                <div class="control">
+                                    <button class="button is-info" type="submit">
+                                        { "Search" }
                                     </button>
                                 </div>
-                                <div class="dropdown-menu" id="dropdown-menu" role="menu">
-                                    <div class="dropdown-content">
-                                        <a
-                                            href={format!("{}/onix/{}", THOTH_API, &w.work_id)}
-                                            class="dropdown-item"
-                                        >
-                                        {"ONIX"}
-                                        </a>
-                                    </div>
-                                </div>
                             </div>
-                        </div>
-                    </nav>
-                </div>
-            </article>
-        </div>
+                        </form>
+                    </div>
+                </nav>
+                {
+                    match self.fetch_data.as_ref().state() {
+                        FetchState::NotFetching(_) => {
+                            html! {<Reloader onclick=self.link.callback(|_| Msg::GetData)/>}
+                        }
+                        FetchState::Fetching(_) => html! {<Loader/>},
+                        FetchState::Fetched(_body) => html! {
+                            { for self.data.iter().map(|w| w.as_catalogue_box()) }
+                        },
+                        FetchState::Failed(_, err) => html! {&err},
+                    }
+                }
+            </div>
+        }
     }
 }