+++ thoth-app/src/component/serieses.rs
@@ -1,150 +1,21 @@
-use yew::html;
-use yew::prelude::*;
-use yew::ComponentLink;
-use yew_router::agent::RouteAgentDispatcher;
-use yew_router::agent::RouteRequest;
-use yew_router::prelude::RouterAnchor;
-use yew_router::route::Route;
-use yewtil::fetch::FetchAction;
-use yewtil::fetch::FetchState;
-use yewtil::future::LinkFuture;
-
-use crate::component::utils::Loader;
-use crate::component::utils::Reloader;
 use crate::models::series::serieses_query::FetchActionSerieses;
 use crate::models::series::serieses_query::FetchSerieses;
+use crate::models::series::serieses_query::SeriesesRequest;
+use crate::models::series::serieses_query::SeriesesRequestBody;
+use crate::models::series::serieses_query::Variables;
 use crate::models::series::Series;
-use crate::route::AdminRoute;
-use crate::route::AppRoute;
 
-pub struct SeriesesComponent {
-    markdown: FetchSerieses,
-    link: ComponentLink<Self>,
-    router: RouteAgentDispatcher<()>,
-}
-
-pub enum Msg {
-    SetMarkdownFetchState(FetchActionSerieses),
-    GetMarkdown,
-    ChangeRoute(AppRoute),
-}
-
-impl Component for SeriesesComponent {
-    type Message = Msg;
-    type Properties = ();
-
-    fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
-        let router = RouteAgentDispatcher::new();
-
-        SeriesesComponent {
-            markdown: Default::default(),
-            link,
-            router,
-        }
-    }
-
-    fn rendered(&mut self, first_render: bool) {
-        if first_render {
-            self.link
-                .send_future(self.markdown.fetch(Msg::SetMarkdownFetchState));
-            self.link
-                .send_message(Msg::SetMarkdownFetchState(FetchAction::Fetching));
-        }
-    }
-
-    fn update(&mut self, msg: Self::Message) -> ShouldRender {
-        match msg {
-            Msg::SetMarkdownFetchState(fetch_state) => {
-                self.markdown.apply(fetch_state);
-                true
-            }
-            Msg::GetMarkdown => {
-                self.link
-                    .send_future(self.markdown.fetch(Msg::SetMarkdownFetchState));
-                self.link
-                    .send_message(Msg::SetMarkdownFetchState(FetchAction::Fetching));
-                false
-            }
-            Msg::ChangeRoute(r) => {
-                let route = Route::from(r);
-                self.router.send(RouteRequest::ChangeRoute(route));
-                false
-            }
-        }
-    }
-
-    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-        false
-    }
-
-    fn view(&self) -> Html {
-        match self.markdown.as_ref().state() {
-            FetchState::NotFetching(_) => {
-                html! {<Reloader onclick=self.link.callback(|_| Msg::GetMarkdown)/>}
-            }
-            FetchState::Fetching(_) => html! {<Loader/>},
-            FetchState::Fetched(body) => html! {
-                <>
-                    <nav class="level">
-                        <div class="level-left">
-                            <div class="level-item">
-                                <p class="subtitle is-5">
-                                    <strong>{ body.data.serieses.iter().count() }</strong> { " series" }
-                                </p>
-                            </div>
-                        </div>
-                        <div class="level-right">
-                            <p class="level-item">
-                                <RouterAnchor<AppRoute>
-                                    classes="button is-success"
-                                    route=AppRoute::Admin(AdminRoute::NewSeries)
-                                >
-                                    {"New"}
-                                </  RouterAnchor<AppRoute>>
-                            </p>
-                        </div>
-                    </nav>
-                    <table class="table is-fullwidth is-hoverable">
-                        <thead>
-                            <tr>
-                                <th>{ "ID" }</th>
-                                <th>{ "Series" }</th>
-                                <th>{ "Series Type" }</th>
-                                <th>{ "ISSN Print" }</th>
-                                <th>{ "ISSN Digital" }</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            { for body.data.serieses.iter().map(|p| self.render_series(p)) }
-                        </tbody>
-                    </table>
-                </>
-            },
-            FetchState::Failed(_, err) => html! {&err},
-        }
-    }
-}
-
-impl SeriesesComponent {
-    fn change_route(&self, app_route: AppRoute) -> Callback<MouseEvent> {
-        self.link.callback(move |_| {
-            let route = app_route.clone();
-            Msg::ChangeRoute(route)
-        })
-    }
-
-    fn render_series(&self, p: &Series) -> Html {
-        html! {
-            <tr
-                class="row"
-                onclick=&self.change_route(AppRoute::Admin(AdminRoute::Series(p.series_id.clone())))
-            >
-                <td>{&p.series_id}</td>
-                <td>{&p.series_name}</td>
-                <td>{&p.series_type}</td>
-                <td>{&p.issn_print}</td>
-                <td>{&p.issn_digital}</td>
-            </tr>
-        }
-    }
+pagination_component! {
+    SeriesesComponent,
+    Series,
+    serieses,
+    series_count,
+    SeriesesRequest,
+    FetchActionSerieses,
+    FetchSerieses,
+    SeriesesRequestBody,
+    Variables,
+    SEARCH_SERIESES,
+    PAGINATION_COUNT_SERIESES,
+    vec!["ID".to_string(), "Series".to_string(), "SeriesType".to_string(), "ISSNPrint".to_string(), "ISSNDigital".to_string()]
 }