+++ thoth-app/src/models/publisher/publishers_query.rs
@@ -4,13 +4,14 @@ use serde::Serialize;
 use super::Publisher;
 
 const PUBLISHERS_QUERY: &str = "
-    {
-        publishers(limit: 9999) {
+    query PublishersQuery($limit: Int, $offset: Int, $filter: String) {
+        publishers(limit: $limit, offset: $offset, filter: $filter) {
             publisherId
             publisherName
             publisherShortname
             publisherUrl
         }
+        publisherCount(filter: $filter)
     }
 ";
 
@@ -26,9 +27,16 @@ graphql_query_builder! {
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
-pub struct Variables {}
+#[serde(rename_all = "camelCase")]
+pub struct Variables {
+    pub limit: Option<i32>,
+    pub offset: Option<i32>,
+    pub filter: Option<String>,
+}
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
 pub struct PublishersResponseData {
     pub publishers: Vec<Publisher>,
+    pub publisher_count: i32,
 }