+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to thoth will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [[0.2.3]](https://github.com/thoth-pub/thoth/releases/tag/v0.2.3) - 2020-11-06
+### Added
+  - Implemented pagination in all admin components
+  - Implemented pagination in catalogue
+
 ## [[0.2.2]](https://github.com/thoth-pub/thoth/releases/tag/v0.2.2) - 2020-11-03
 ### Changed
   - Set `THOTH_API` on build via docker